requirements.txt
Pillow
selenium
webdriver_manager
undetected-chromedriver
dnspython
beautifulsoup4
requests
